Propagate metadata from predictions into curated files (#117)
Closes #114

This PR adds 3 extra columns to the curated files (correct, incorrect,
unsure):

1. `prediction_type` - e.g., `lexical`
2. `prediction_source` - e.g., this has the URL or name of the script
that made the data
3. `prediction_confidence` - e.g., the confidence assigned by the script
that produced predictions

It also rewires various parts of the I/O and web application to make
sure that prediction metadata gets forwarded to the manually curated
files going forward. This PR does not attempt to reverse-engineer this
metadata for existing manual curations.