Fix incorrectly encoded manual curations
I think during the process when I was adding the extra provenance columns to the mappings.tsv, I accepted some curations where the prediction type (i.e., lexical) was rolled over into the manual curation column. This commit fixes the 74 instances where this happened. This is the script I wrote to fix this:

```python
import pandas as pd

def main():
    df = pd.read_csv("mappings.tsv", sep='\t')
    idx = df['type'] == 'lexical'
    df.loc[idx, 'type'] = "semapv:ManualMatchingCuration"
    df.loc[idx, 'prediction_type'] = "semapv:LexicalMatching"
    df.to_csv("mappings.tsv", sep='\t', index=False)

if __name__ == '__main__':
    main()
```