Improve redundancy filtering (#141)
Before, redundant mapping filtering just checked that the tuple was the
same. Now, it groups by the canonical tuple and uses a key for picking
the "best" mapping. For now, this is a simple function. Later, it can be
extended to take into account confidence in the curator, prediction
methodology, or potentially extending the data model to keep track of
the date of prediction/curation and just keep the earliest.