Add generic contribution workflow to OBO flat files (#149)
This PR adds a script that can be used to contribute Biomappings
curations to a given OBO flat file. It is careful to check the file for
pre-existing curations and not to add duplicates. It does its best to
maintain a deterministic order for where it adds new xrefs within a
given stanza.

Future: use more detailed relationship information (e.g.,
`skos:exactMatch`)