Map Iconclass to ChEBI, OBI, and CHMO (#205)
This PR adds three curation scenarios:

1. Embedding-based matches between Iconclass and CHMO
2. Embedding-based matches between Iconclass and OBI
3. NER-based matches between Iconclass and ChEBI (e.g., find ChEBI
entities appearing as part of the names in Iconclass). This is because
the embeddings are very far off, and most of the terms don't have exact
matches.

In all three scenarios, I used skos:relatedMatch because Iconclass is
about the depiction of things, and CHMO/OBI/ChEBI are about the things
themselves.

In general, the predictions are quite poor quality so I only included
positive and negative curations but not the potentially thousands of low
quality predictions. I focused ChEBI curation around entities that had
metals or atoms in them. For CHMO and OBI, I was a bit more exhaustive.