Refactor prediction appending and add CLI (#224)
This adds a CLI `biomappings predict mesh maxo` anything can be added
quickly in an ad-hoc way.

This also improves the predictions append/write/lint functions, then
removes some old code and CLI endpoints that are now subsumed by
`biomappings lint`

It finally restructures some of the code that works on mappings in
preparation for upstreaming it.