Upstream implementation to SSSOM Curator (#232)
This PR externalizes the code in Biomappings to SSSOM-Curator
(https://github.com/cthoyt/sssom-curator), which generalizes a lot of
the functionality of Biomappings and re-brands it as domain-agnostic. I
haven't decided yet where to house this generalized functionality yet,
so it lives in my personal account for the moment.

This PR doesn't affect the CLI nor the functions exposed from
Biomappings, though, the last few updates have been moving towards a
reusable object model for semantic mappings from ssson-pydantic, so some
parts will lose backwards compatibility because of that. This will come
with a minor version bump.

This was motivated by NFDI use cases, where people outside of the
biomedical domain are more or less consistently put off by the
nomenclature of software components/databases in the biopragmatics
stack.